# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CU Faculty Backend API
  version: "1.0.0"
  description: >
    REST API for the CU Faculty student hub. All endpoints under `/api` require a
    valid `bff.cookie` session cookie, which is validated against the Central
    University (CU) upstream on every request. The authenticated CU student id is
    used as the local user id. The only unauthenticated endpoint is `/health`.


    List endpoints are paginated and return a `{ data, total, limit, offset }`
    envelope. Pagination is controlled with the `limit` (default 20, max 100) and
    `offset` (default 0) query parameters.
security:
  - cookieAuth: []
...
paths:
...
  /api/me/contacts:
    get:
      tags: [Me]
      summary: List my contacts
      operationId: getMyContacts
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged contacts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContactPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Me]
      summary: Add a contact
      description: >
        Links the current user to another user in both directions. The reverse
        link is created without a note.
      operationId: createContact
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateContactRequest" }
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }
        "400":
          description: Bad request or attempt to add yourself
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Referenced contact user does not exist
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Contact already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/me/contacts/{contactId}:
    put:
      tags: [Me]
      summary: Update a contact note
      operationId: updateContact
      parameters:
        - $ref: "#/components/parameters/ContactId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateContactRequest" }
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Me]
      summary: Delete a contact
      operationId: deleteContact
      parameters:
        - $ref: "#/components/parameters/ContactId"
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/me/announcements:
    get:
      tags: [Me]
      summary: List announcements I authored
      operationId: getMyAnnouncements
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged announcements
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnnouncementPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/announcements:
    get:
      tags: [Announcements]
      summary: List announcements
      operationId: listAnnouncements
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged announcements
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnnouncementPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Announcements]
      summary: Create an announcement
      operationId: createAnnouncement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAnnouncementRequest" }
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/announcements/{id}:
    get:
      tags: [Announcements]
      summary: Get an announcement
      operationId: getAnnouncement
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [Announcements]
      summary: Update an announcement (author only)
      operationId: updateAnnouncement
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAnnouncementRequest" }
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Announcements]
      summary: Delete an announcement (author only)
      operationId: deleteAnnouncement
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/announcements/{id}/responses:
    get:
      tags: [Announcements]
      summary: List responders to an announcement
      operationId: getAnnouncementResponders
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged responders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Announcements]
      summary: Respond to an announcement
      description: Idempotent. Fails if the announcement is missing or archived.
      operationId: respondToAnnouncement
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "201":
          description: Announcement responded to
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Announcements]
      summary: Remove my response to an announcement
      operationId: deleteMyAnnouncementResponse
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204": { description: Removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: bff.cookie
      description: CU session cookie, validated against the CU upstream on every request.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size (default 20, clamped to a maximum of 100).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip (default 0).
      schema: { type: integer, minimum: 0, default: 0 }
...
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ContactId:
      name: contactId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or rejected session cookie
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: >
        Normalized error envelope returned by every error response. "error" is a
        human-readable message (not for programmatic use); "code" is a stable,
        machine-readable identifier clients may branch on.
      properties:
        error: { type: string, example: "announcement not found" }
        code:
          type: string
          example: not_found
          enum:
            - bad_request
            - validation_error
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - already_exists
            - invalid_reference
            - payload_too_large
            - upstream_error
            - unavailable
            - internal_error
      required: [error, code]
    PageMeta:
      type: object
      properties:
        total: { type: integer, description: Total number of matching items }
        limit: { type: integer }
        offset: { type: integer }
      required: [total, limit, offset]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        photoUrl:
          type: string
          nullable: true
          description: Presigned, time-limited download URL for the user's photo.
        firstName: { type: string, maxLength: 31 }
        lastName: { type: string, maxLength: 31 }
        bio: { type: string, nullable: true, maxLength: 255 }
        birthdate: { type: string, format: date, example: "2000-01-31" }
        speciality: { type: string, nullable: true, maxLength: 63 }
        status: { type: string, nullable: true, maxLength: 127, description: Resolved status content }
        role: { type: string, enum: [user, admin] }
      required: [id, firstName, lastName, birthdate, role]
...
    Contact:
      type: object
      properties:
        note: { type: string, nullable: true, maxLength: 255 }
        user: { $ref: "#/components/schemas/User" }
      required: [user]
    CreateContactRequest:
      type: object
      properties:
        contactId: { type: string, format: uuid }
        note: { type: string, nullable: true, maxLength: 255 }
      required: [contactId]
    UpdateContactRequest:
      type: object
      properties:
        note: { type: string, nullable: true, maxLength: 255 }
...
    Announcement:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 63 }
        content: { type: string, maxLength: 255 }
        isArchived: { type: boolean }
        createdAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/User" }
      required: [id, title, content, isArchived, createdAt, author]
    CreateAnnouncementRequest:
      type: object
      properties:
        title: { type: string, maxLength: 63 }
        content: { type: string, maxLength: 255 }
      required: [title, content]
    UpdateAnnouncementRequest:
      type: object
      properties:
        title: { type: string, maxLength: 63 }
        content: { type: string, maxLength: 255 }
        isArchived: { type: boolean }
      required: [title, content]
...
    UserPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/User" } }
          required: [data]
...
    ContactPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/Contact" } }
          required: [data]
...
    AnnouncementPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/Announcement" } }
          required: [data]
...